Log copy errors
The following commit logs all copy errors when copying from writers to
readers. The should help with understanding why some connections can end
up with packets that are not delivered, but they should be.

The error is not surfaced if there was one trying to stream the data.

We can also help diagnose issues by also checking the error when
closing the connection.